insertions_service/requirements.txt
Django
Djongo
pymongo==3.12.3
pytz
Pillow
typing-extensions
djangorestframework
django-cors-headers